Group all license text specs into one file
